+++ package.json
@@ -12,14 +12,19 @@
 		"jsonwebtoken": "^9.0.2",
 		"next": "14.2.3",
 		"next-connect": "^1.0.0",
-		"nodemailer": "^6.9.13",
+		"plyr": "^3.7.8",
+		"plyr-react": "^5.3.0",
 		"preline": "^2.1.0",
 		"qrcode": "^1.5.3",
 		"react": "^18",
 		"react-dom": "^18",
+		"react-hot-toast": "^2.4.1",
 		"react-icons": "^5.2.1",
+		"react-player": "^2.16.0",
 		"sqlite3": "^5.1.7",
 		"string-strip-html": "^13.4.8",
+		"sweetalert2": "^11.11.0",
+		"sweetalert2-react-content": "^5.0.7",
 		"swiper": "^11.1.3",
 		"ua-parser-js": "^1.0.37",
 		"zod": "^3.23.8"
