+++ src/components/courseLesson/unit/NavCh.jsx
@@ -1,0 +1,89 @@
+import React from 'react';
+import VideoIframe from '../../coursevideo/content';
+import Chapter from './chapter';
+import ChapterVideo from './chapter';
+
+export default function NavCh() {
+	return (
+		<>
+			
+			<Chapter />
+			<ChapterVideo />
+			{/* 輪播 */}
+			<div
+				data-hs-carousel='{"loadingClasses": "opacity-0"}'
+				className="relative">
+				<div className="hs-carousel relative overflow-hidden min-h-96 bg-white rounded-lg">
+					<div className="hs-carousel-body absolute top-0 bottom-0 start-0 flex flex-nowrap transition-transform duration-700 opacity-0">
+						<div className="hs-carousel-slide">
+							<div className="pl-30 flex justify-start   h-full bg-gray-100 ">
+								<span className="self-center text-4xl text-gray-800 transition duration-700">
+									<VideoIframe />
+								</span>
+							</div>
+						</div>
+						<div className="hs-carousel-slide">
+							<div className="flex justify-center h-full bg-gray-200 p-6">
+								<span className="self-center text-4xl text-gray-800 transition duration-700"></span>
+							</div>
+						</div>
+						<div className="hs-carousel-slide">
+							<div className="flex justify-center h-full bg-gray-300 p-6">
+								<span className="self-center text-4xl text-gray-800 transition duration-700"></span>
+							</div>
+						</div>
+					</div>
+				</div>
+				<button
+					type="button"
+					className="hs-carousel-prev hs-carousel:disabled:opacity-50 disabled:pointer-events-none absolute inset-y-0 start-0 inline-flex justify-center items-center w-[46px] h-full text-gray-800 hover:bg-gray-800/10 rounded-s-lg">
+					<span
+						className="text-2xl"
+						aria-hidden="true">
+						<svg
+							className="flex-shrink-0 size-5"
+							xmlns="http://www.w3.org/2000/svg"
+							width={24}
+							height={24}
+							viewBox="0 0 24 24"
+							fill="none"
+							stroke="currentColor"
+							strokeWidth={2}
+							strokeLinecap="round"
+							strokeLinejoin="round">
+							<path d="m15 18-6-6 6-6" />
+						</svg>
+					</span>
+					<span className="sr-only">Previous</span>
+				</button>
+				<button
+					type="button"
+					className="hs-carousel-next hs-carousel:disabled:opacity-50 disabled:pointer-events-none absolute inset-y-0 end-0 inline-flex justify-center items-center w-[46px] h-full text-gray-800 hover:bg-gray-800/10 rounded-e-lg">
+					<span className="sr-only">Next</span>
+					<span
+						className="text-2xl"
+						aria-hidden="true">
+						<svg
+							className="flex-shrink-0 size-5"
+							xmlns="http://www.w3.org/2000/svg"
+							width={24}
+							height={24}
+							viewBox="0 0 24 24"
+							fill="none"
+							stroke="currentColor"
+							strokeWidth={2}
+							strokeLinecap="round"
+							strokeLinejoin="round">
+							<path d="m9 18 6-6-6-6" />
+						</svg>
+					</span>
+				</button>
+				<div className="hs-carousel-pagination flex justify-center absolute bottom-3 start-0 end-0 space-x-2">
+					<span className="hs-carousel-active:bg-blue-700 hs-carousel-active:border-blue-700 size-3 border border-gray-400 rounded-full cursor-pointer" />
+					<span className="hs-carousel-active:bg-blue-700 hs-carousel-active:border-blue-700 size-3 border border-gray-400 rounded-full cursor-pointer" />
+					<span className="hs-carousel-active:bg-blue-700 hs-carousel-active:border-blue-700 size-3 border border-gray-400 rounded-full cursor-pointer" />
+				</div>
+			</div>
+		</>
+	);
+}
