--- src/components/video.tsx
@@ -1,49 +1,0 @@
-import React from 'react';
-
-// function VideoIframe() {
-// 	return (
-// 		<div>
-// 			<iframe
-//           className="rounded-xl md:w-600 sm:h-366"
-//           width="380"
-//       height="266px"
-//       src={props.src}
-//           allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" referrerPolicy="strict-origin-when-cross-origin"
-//           allowFullScreen
-          
-//             ></iframe>
-
-// 			<iframe
-// 				width="500"
-// 				height="300"
-// 				src="https://www.youtube.com/embed/bLynTAwa40g"
-// 				title="兩分鐘帶您感受冬日的氣氛 ♪【Winter Reflections l Special Occasions】♪"
-// 				frameBorder="0"
-// 				allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share"
-// 				referrerPolicy="strict-origin-when-cross-origin"
-// 				allowFullScreen></iframe>
-// 		</div>
-// 	);
-// }
-
-// export default VideoIframe;
-
-// import React from 'react';
-
-export default function Hotcard() {
-	return (
-		<>
-			{/* <video
-				className="h-full w-full rounded-lg"
-				controls
-				autoPlay
-				muted>
-				<source
-					src="https://videos.pexels.com/video-files/7989707/7989707-hd_1920_1080_25fps.mp4"
-					type="video/mp4"
-				/>
-				Your browser does not support the video tag.
-			</video> */}
-		</>
-	);
-}
