--- src/pages/api/wishlist/[sn].js
@@ -1,8 +1,0 @@
-import connection from '@/handlers/sqlite3';
-
-export default async function handler(req, res) {
-    const sn = req.query.sn;
-    const [results] = await connection.execute('SELECT * FROM Course WHERE SN =?', [sn]);
-
-    res.status(200).json(results[0]);
-}
