+++ src/pages/course/test.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react';
 
-import CardFirst from '../../components/card/cardfirst';
-import Card from '../../components/card/Card';
-
+// import CardFirst from '../../components/card/cardfirst';
+// import Card from '../../components/card/Card';
+// import Bookmark from '../../components/card/bookmark';
 // import CourseList from './testdata';
 
 
@@ -10,7 +10,7 @@ export default function Test() {
     return (
         <>
             {/* <div className="flex flex-wrap space-x-3">  <CardFirst /></div> */}
-            <Card/>
+            {/* <Bookmark /> */}
           
         </>
     )
